updates object classes | disabled superclass for hibernate
