+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=update
 spring.datasource.url=jdbc:derby:studentxdb;create=true
 logging.level.org.hibernate.SQL=DEBUG
 spring.jpa.show-sql=true
